fusillade authorization operation
